+++ src/commands/index.ts
@@ -5,3 +5,4 @@ export * as page from "./page";
 export * as invoke from "./invoke";
 export * as lorem from "./lorem";
 export * as sort from "./sort";
+export * as emoji from "./emoji";